Factor out common types from NodeViewHolderGenerators and NodeViewSynchronizerGenerators
